// types/listingTypes.ts
export interface ListingType {
    id: string
    name: string
    image: string
    description: string
    titletheory1: string
    theory1: string
    imageth1a: string
    imageth1b: string
    imageth1c: string
    titletheory2: string
    theory2: string
    imageth2a: string
    imageth2b: string
    imageth2c: string
    titletheory3: string
    theory3: string
    imageth3a: string
    imageth3b: string
    imageth3c: string
    titletheory4: string
    theory4: string
    imageth4a: string
    imageth4b: string
    imageth4c: string
    titletheory5: string
    theory5: string
    imageth5a: string
    imageth5b: string
    imageth5c: string
    titletheory6: string
    theory6: string
    imageth6a: string
    imageth6b: string
    imageth6c: string
    titletheory7: string
    theory7: string
    imageth7a: string
    imageth7b: string
    imageth7c: string
    titletheory8: string
    theory8: string
    imageth8a: string
    imageth8b: string
    imageth8c: string
    titletheory9: string
    theory9: string
    imageth9a: string
    imageth9b: string
    imageth9c: string
    titletheory10: string
    theory10: string
    imageth10a: string
    imageth10b: string
    imageth10c: string
    titletheory11: string
    theory11: string
    imageth11a: string
    imageth11b: string
    imageth11c: string
    titletheory12: string
    theory12: string
    imageth12a: string
    imageth12b: string
    imageth12c: string
    titletheory13: string
    theory13: string
    imageth13a: string
    imageth13b: string
    imageth13c: string
    titletheory14: string
    theory14: string
    imageth14a: string
    imageth14b: string
    imageth14c: string
    titletheory15: string
    theory15: string
    imageth15a: string
    imageth15b: string
    imageth15c: string
    titletheory16: string
    theory16: string
    imageth16a: string
    imageth16b: string
    imageth16c: string
    titletheory17: string
    theory17: string
    imageth17a: string
    imageth17b: string
    imageth17c: string
    titletheory18: string
    theory18: string
    imageth18a: string
    imageth18b: string
    imageth18c: string
    titletheory19: string
    theory19: string
    imageth19a: string
    imageth19b: string
    imageth19c: string
    titletheory20: string
    theory20: string
    imageth20a: string
    imageth20b: string
    imageth20c: string
    titletheory21: string
    theory21: string
    imageth21a: string
    imageth21b: string
    imageth21c: string
    titletheory22: string
    theory22: string
    imageth22a: string
    imageth22b: string
    imageth22c: string
    titletheory23: string
    theory23: string
    imageth23a: string
    imageth23b: string
    imageth23c: string
    titletheory24: string
    theory24: string
    imageth24a: string
    imageth24b: string
    imageth24c: string
    titletheory25: string
    theory25: string
    imageth25a: string
    imageth25b: string
    imageth25c: string
    titletheory26: string
    theory26: string
    imageth26a: string
    imageth26b: string
    imageth26c: string
    titletheory27: string
    theory27: string
    imageth27a: string
    imageth27b: string
    imageth27c: string
    titletheory28: string
    theory28: string
    imageth28a: string
    imageth28b: string
    imageth28c: string
    titletheory29: string
    theory29: string
    imageth29a: string
    imageth29b: string
    imageth29c: string
    titletheory30: string
    theory30: string
    imageth30a: string
    imageth30b: string
    imageth30c: string
    titletheory31: string
    theory31: string
    imageth31a: string
    imageth31b: string
    imageth31c: string
    titletheory32: string
    theory32: string
    imageth32a: string
    imageth32b: string
    imageth32c: string
    titletheory33: string
    theory33: string
    imageth33a: string
    imageth33b: string
    imageth33c: string
    titletheory34: string
    theory34: string
    imageth34a: string
    imageth34b: string
    imageth34c: string
    titletheory35: string
    theory35: string
    imageth35a: string
    imageth35b: string
    imageth35c: string
    titletheory36: string
    theory36: string
    imageth36a: string
    imageth36b: string
    imageth36c: string
    titletheory37: string
    theory37: string
    imageth37a: string
    imageth37b: string
    imageth37c: string
    titletheory38: string
    theory38: string
    imageth38a: string
    imageth38b: string
    imageth38c: string
    titletheory39: string
    theory39: string
    imageth39a: string
    imageth39b: string
    imageth39c: string
    titletheory40: string
    theory40: string
    imageth40a: string
    imageth40b: string
    imageth40c: string
    section1: string
    section2: string
    section3: string
    category: string
    videos?: string[]; 
  }